webhook-service/requirements.txt
Flask==3.0.3
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.0
google-cloud-storage==2.10.0
google-api-python-client==2.108.0
google-auth==2.23.4
weasyprint==66.0 